Lower affinity of PKA to D1R
Gardner et al., 2011, states that the affinity of D1R to dopamine for
phosphorylation is roughly 200 nM and it is 10 x higher than the
affinity of D1R for DA. I lowered the affinity for phosphorylation
because D1R affinity for DA is much lower than 20 nM (3 uM roughly).
Also D1R desensitizes to quickly and cAMP timecourse does not match the
experimental from Muntean et al, 2018.